# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DeployMonster API
  version: 1.4.0
  description: |
    Self-hosted PaaS API for managing applications, deployments, domains,
    secrets, teams, and infrastructure. Single binary, modular monolith
    with event-driven architecture.
  contact:
    name: DeployMonster
    url: https://deploy.monster
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8443
    description: Local development server
  - url: https://{host}
    description: Production server
    variables:
      host:
        default: deploy.monster
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/team/audit-log:
    get:
      operationId: getAuditLog
      tags: [Team]
      summary: Get team audit log
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        user_id:
                          type: string
                        action:
                          type: string
                        resource:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        details:
                          type: object

  # ── Auth (additional) ──────────────────────────────────
...
  /api/v1/team/members/{id}:
    delete:
      operationId: removeTeamMember
      tags: [Team]
      summary: Remove a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team member removed

  # ── Events & Search ──────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated from admin panel
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
...